+++ src/multitenant_sdk.js
@@ -1,7 +1,10 @@
 import _ from 'lodash';
-import { fnPath as urlPathToFnPath, trimEndSlash, formData } from './utils';
+import { fnPath as urlPathToFnPath, trimEndSlash } from './utils';
+import { authApi as authApiEndpoints } from './endpoints';
 import AddMultitenantAuthTokenResponse from './interceptors/add_multitenant_auth_token_response';
 import SaveToken from './interceptors/save_token';
+import RetryWithRefreshToken from './interceptors/retry_with_refresh_token';
+import FetchAuthTokenFromStore from './interceptors/fetch_auth_token_from_store';
 import AddMultitenantClientSecretTokenToCtx from './interceptors/add_multitenant_client_secret_token_to_ctx';
 import AddMultitenantClientSecretToParams from './interceptors/add_multitenant_client_secret_to_params';
 import AddMultitenantTokenExchangeParams from './interceptors/add_multitenant_token_exchange_params';
@@ -15,17 +18,14 @@ import createSdkFnContextRunner from './sdk_context_runner';
 import memoryStore from './memory_store';
 import contextRunner from './context_runner';
 import { isBrowser } from './runtime';
+import * as authTokenContextRunner from './auth_token_context_runner';
+import { apis, defaultSdkConfig } from './api_config';
 
 /* eslint-disable class-methods-use-this */
 
-const defaultSdkConfig = {
+const multitenantSdkConfig = {
   hostname: null,
   multitenantClientSecret: null,
-  baseUrl: 'https://flex-api.sharetribe.com',
-  adapter: null,
-  version: 'v1',
-  httpAgent: null,
-  httpsAgent: null,
 };
 
 const multitenantAuthApi = [
@@ -43,25 +43,24 @@ const multitenantAuthApi = [
   },
 ];
 
-const apis = {
-  auth: ({ baseUrl, version, adapter, httpAgent, httpsAgent }) => ({
-    headers: {
-      'Content-Type': 'application/x-www-form-urlencoded',
-      Accept: 'application/json',
-    },
-    baseURL: `${baseUrl}/${version}/`,
-    transformRequest: [data => formData(data)],
-    adapter,
-    httpAgent,
-    httpsAgent,
-  }),
-};
-
 const tokenInterceptors = authApiEndpointInterceptors => [
   new FormatHttpResponse(),
   new FormatMultitenantHttpResponse(),
   new AddMultitenantClientSecretTokenToCtx(),
   new AddMultitenantClientSecretToParams(),
+  new SaveToken(),
+  new AddMultitenantAuthTokenResponse(),
+  ..._.get(authApiEndpointInterceptors, 'token'),
+];
+
+const tokenExchangeInterceptors = authApiEndpointInterceptors => [
+  new FormatHttpResponse(),
+  new FormatMultitenantHttpResponse(),
+  new FetchAuthTokenFromStore(),
+  new RetryWithRefreshToken(),
+  new AddMultitenantClientSecretTokenToCtx(),
+  new AddMultitenantClientSecretToParams(),
+  new AddMultitenantTokenExchangeParams(),
   new SaveToken(),
   new AddMultitenantAuthTokenResponse(),
   ..._.get(authApiEndpointInterceptors, 'token'),
@@ -80,11 +79,21 @@ const tokenAndClientDataInterceptor = authApiEndpointInterceptors => ({
     return Promise.resolve()
       .then(tokenStore.getToken)
       .then(storedToken => {
-        // If there's a token with any access, it's only necessary
-        // to fetch the client data. Else, we request a token and
-        // the response will also contain the client data.
-        // We don't need to distinguish between token scopes.
+        // If there's a token with any access, it's only necessary to fetch the
+        // client data. Else, we request a token using
+        // multitenant_client_credentials grant type.
+        //
+        // The returned token response contains the access token AND client
+        // data. This is an _optimization_ so that we can get client data AND a
+        // valid (anonymous) token with one request instead of two.
+        //
+        // We don't need to distinguish between token scopes. We don't check for
+        // token validity. If the token is invalid/expired, clientData request
+        // will go through because it doesn't require access token. The expired
+        // token will be refreshed by the "normal SDK". Everything works but we
+        // just didn't gain the advance from the optimization.
         if (storedToken) {
+          // Request client data only
           return contextRunner(clientDataInterceptors(authApiEndpointInterceptors))(ctx).then(
             newCtx => {
               const { res } = newCtx;
@@ -101,6 +110,7 @@ const tokenAndClientDataInterceptor = authApiEndpointInterceptors => ({
           );
         }
 
+        // Request token (which includes client data)
         return contextRunner(tokenInterceptors(authApiEndpointInterceptors))({
           ...ctx,
           params: { grant_type: 'multitenant_client_credentials' },
@@ -137,10 +147,7 @@ const authApiSdkFns = (authApiEndpointInterceptors, ctx) => [
     path: 'tokenExchange',
     fn: createAuthApiSdkFn({
       ctx,
-      interceptors: [
-        new AddMultitenantTokenExchangeParams(),
-        ...tokenInterceptors(authApiEndpointInterceptors),
-      ],
+      interceptors: tokenExchangeInterceptors(authApiEndpointInterceptors),
     }),
   },
   {
@@ -181,11 +188,13 @@ const validateSdkConfig = sdkConfig => {
 };
 
 const createAuthApiEndpointInterceptors = httpOpts =>
-  // Create `endpointInterceptors` object, which is object
-  // containing interceptors for all defined endpoints.
-  // This object can be passed to other interceptors in the interceptor context so they
-  // are able to do API calls (e.g. authentication interceptors)
-  //
+  authApiEndpoints.reduce((acc, { path, method }) => {
+    const fnPath = urlPathToFnPath(path);
+    const url = `auth/${path}`;
+    return _.set(acc, fnPath, [endpointRequest({ method, url, httpOpts })]);
+  }, {});
+
+const createMultitenantAuthApiEndpointInterceptors = httpOpts =>
   multitenantAuthApi.reduce((acc, { path, method }) => {
     const fnPath = urlPathToFnPath(path);
     const url = `auth/multitenant/${path}`;
@@ -201,26 +210,26 @@ export default class MultitenantSharetribeSdk {
   constructor(userSdkConfig) {
     // Transform and validation SDK configurations
     const sdkConfig = validateSdkConfig(
-      transformSdkConfig({ ...defaultSdkConfig, ...userSdkConfig })
+      transformSdkConfig({ ...defaultSdkConfig, ...multitenantSdkConfig, ...userSdkConfig })
     );
 
     // Instantiate API configs
     const apiConfigs = _.mapValues(apis, apiConfig => apiConfig(sdkConfig));
     const authApiEndpointInterceptors = createAuthApiEndpointInterceptors(apiConfigs.auth);
+    const multitenantAuthApiEndpointInterceptors = createMultitenantAuthApiEndpointInterceptors(
+      apiConfigs.auth
+    );
 
-    const allEndpointInterceptors = {
-      auth: authApiEndpointInterceptors,
-    };
-
-    const ctx = {
-      endpointInterceptors: allEndpointInterceptors,
+    let ctx = {
       multitenantClientSecret: sdkConfig.multitenantClientSecret,
       hostname: sdkConfig.hostname,
       tokenStore: sdkConfig.tokenStore,
     };
 
+    ctx = authTokenContextRunner.setAuthTokenInterceptors(ctx, authApiEndpointInterceptors.token);
+
     // Assign SDK functions to 'this'
-    authApiSdkFns(authApiEndpointInterceptors, ctx).forEach(({ path, fn }) =>
+    authApiSdkFns(multitenantAuthApiEndpointInterceptors, ctx).forEach(({ path, fn }) =>
       _.set(this, path, fn)
     );
   }